backend/fastapi_app/requirements.txt
# FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# AI/ML
openai==1.3.7

# HTTP client
httpx==0.25.2
requests==2.31.0

# Async support
aiofiles==23.2.1

# Utilities
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# Security
cryptography==41.0.8
